+++ tests/vertex_buffer.rs
@@ -4,6 +4,7 @@ extern crate glutin;
 extern crate glium;
 
 use glium::Surface;
+use std::default::Default;
 
 mod support;
 
@@ -360,6 +361,174 @@ fn zero_sized_vertex_buffer() {
     implement_vertex!(Vertex, field1, field2);
 
     glium::VertexBuffer::new(&display, Vec::<Vertex>::new());
+
+    display.assert_no_error();
+}
+
+#[test]
+fn slice_draw_indices() {
+    #[derive(Copy, Clone)]
+    struct Vertex {
+        position: [f32; 2],
+    }
+
+    implement_vertex!(Vertex, position);
+
+    let display = support::build_display();
+    let program = glium::Program::from_source(&display,
+        "
+            #version 110
+
+            attribute vec2 position;
+
+            void main() {
+                gl_Position = vec4(position, 0.0, 1.0);
+            }
+        ",
+        "
+            #version 110
+
+            void main() {
+                gl_FragColor = vec4(1.0, 0.0, 0.0, 1.0);
+            }
+        ",
+        None).unwrap();
+
+    let vb = glium::VertexBuffer::new(&display, vec![
+        Vertex { position: [-1.0,  1.0] }, Vertex { position: [1.0,  1.0] },
+        Vertex { position: [-1.0, -1.0] }, Vertex { position: [1.0, -1.0] },
+    ]);
+
+    let indices = glium::index::TrianglesList(vec![0u16, 1, 2]);
+    let indices = glium::IndexBuffer::new(&display, indices);
+
+    let texture = support::build_renderable_texture(&display);
+    texture.as_surface().clear_color(0.0, 0.0, 0.0, 0.0);
+    texture.as_surface().draw(vb.slice(1, 3).unwrap(), &indices, &program,
+                &glium::uniforms::EmptyUniforms, &Default::default()).unwrap();
+
+    let data: Vec<Vec<(u8, u8, u8)>> = texture.read();
+    assert_eq!(data.last().unwrap()[0], (0, 0, 0));
+    assert_eq!(data[0].last().unwrap(), &(255, 0, 0));
+
+    display.assert_no_error();
+}
+
+#[test]
+fn slice_draw_noindices() {
+    #[derive(Copy, Clone)]
+    struct Vertex {
+        position: [f32; 2],
+    }
+
+    implement_vertex!(Vertex, position);
+
+    let display = support::build_display();
+    let program = glium::Program::from_source(&display,
+        "
+            #version 110
+
+            attribute vec2 position;
+
+            void main() {
+                gl_Position = vec4(position, 0.0, 1.0);
+            }
+        ",
+        "
+            #version 110
+
+            void main() {
+                gl_FragColor = vec4(1.0, 0.0, 0.0, 1.0);
+            }
+        ",
+        None).unwrap();
+
+    let vb = glium::VertexBuffer::new(&display, vec![
+        Vertex { position: [-1.0,  1.0] }, Vertex { position: [1.0,  1.0] },
+        Vertex { position: [-1.0, -1.0] }, Vertex { position: [1.0, -1.0] },
+    ]);
+
+    let indices = glium::index::NoIndices(glium::index::PrimitiveType::TrianglesList);
+
+    let texture = support::build_renderable_texture(&display);
+    texture.as_surface().clear_color(0.0, 0.0, 0.0, 0.0);
+    texture.as_surface().draw(vb.slice(1, 3).unwrap(), &indices, &program,
+                &glium::uniforms::EmptyUniforms, &Default::default()).unwrap();
+
+    let data: Vec<Vec<(u8, u8, u8)>> = texture.read();
+    assert_eq!(data.last().unwrap()[0], (0, 0, 0));
+    assert_eq!(data[0].last().unwrap(), &(255, 0, 0));
+
+    display.assert_no_error();
+}
+
+#[test]
+fn slice_draw_multiple() {
+    #[derive(Copy, Clone)]
+    struct Vertex {
+        position: [f32; 2],
+    }
+
+    implement_vertex!(Vertex, position);
+
+    #[derive(Copy, Clone)]
+    struct Vertex2 {
+        position2: [f32; 2],
+    }
+
+    implement_vertex!(Vertex2, position2);
+
+    let display = support::build_display();
+    let program = glium::Program::from_source(&display,
+        "
+            #version 110
+
+            attribute vec2 position;
+            attribute vec2 position2;
+
+            void main() {
+                if (position != position2) {
+                    gl_Position = vec4(0.0, 0.0, 0.0, 1.0);
+                } else {
+                    gl_Position = vec4(position, 0.0, 1.0);
+                }
+            }
+        ",
+        "
+            #version 110
+
+            void main() {
+                gl_FragColor = vec4(1.0, 0.0, 0.0, 1.0);
+            }
+        ",
+        None).unwrap();
+
+    // the 3 last elements will be drawn
+    let vb1 = glium::VertexBuffer::new(&display, vec![
+        Vertex { position: [-1.0,  1.0] }, Vertex { position: [-1.0, 1.0] },
+        Vertex { position: [-1.0,  1.0] }, Vertex { position: [1.0,  1.0] },
+        Vertex { position: [-1.0, -1.0] }, Vertex { position: [1.0, -1.0] },
+    ]);
+
+    // the 3 last elements will be drawn
+    let vb2 = glium::VertexBuffer::new(&display, vec![
+        Vertex2 { position2: [-1.0,  1.0] }, Vertex2 { position2: [-1.0, 1.0] },
+        Vertex2 { position2: [-1.0,  1.0] }, Vertex2 { position2: [-1.0, 1.0] },
+        Vertex2 { position2: [-1.0,  1.0] }, Vertex2 { position2: [1.0,  1.0] },
+        Vertex2 { position2: [-1.0, -1.0] }, Vertex2 { position2: [1.0, -1.0] },
+    ]);
+
+    let indices = glium::index::TrianglesList(vec![2u16, 3, 4]);
+    let indices = glium::IndexBuffer::new(&display, indices);
+
+    let texture = support::build_renderable_texture(&display);
+    texture.as_surface().clear_color(0.0, 0.0, 0.0, 0.0);
+    texture.as_surface().draw((vb1.slice(1, 3).unwrap(), vb2.slice(3, 3).unwrap()), &indices,
+                &program, &glium::uniforms::EmptyUniforms, &Default::default()).unwrap();
+
+    let data: Vec<Vec<(u8, u8, u8)>> = texture.read();
+    assert_eq!(data.last().unwrap()[0], (0, 0, 0));
+    assert_eq!(data[0].last().unwrap(), &(255, 0, 0));
 
     display.assert_no_error();
 }