+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+ - Added an `implement_uniforms!` macro similar to `implement_vertex!`.
+
 ## Version 0.2.2 (2015-04-10)
 
  - Added support for backends that don't have vertex array objects (like OpenGL ES 2/WebGL).