+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Fixed drawing with offsets in vertex buffers different than 0 not permitted.
  - Changed transform feedback reflection API to be compatible with what OpenGL 4.4 or ARB_enhanced_layouts allow.
  - `VertexBuffer::new` can now take a slice as parameter.
+ - Revert the fix for sRGB. `GL_FRAMEBUFFER_SRGB` is no longer enabled.
 
 ## Version 0.2.2 (2015-04-10)